refactor: move AggregateRating to separate file and reuse in ChannelInfo and ProductGrade
