+++ library/src/main/java/com/etrusted/android/trustbadge/library/model/ChannelInfo.kt
@@ -25,12 +25,9 @@
 
 package com.etrusted.android.trustbadge.library.model
 
-import com.etrusted.android.trustbadge.library.common.internal.fromString
 import org.json.JSONObject
-import java.util.Date
-import com.etrusted.android.trustbadge.library.model.ChannelInfo.AggregateRating.AggregateRatingDistribution
-import com.etrusted.android.trustbadge.library.model.ChannelInfo.AggregateRating.AggregateRatingPeriod
-import com.etrusted.android.trustbadge.library.model.ChannelInfo.AggregateRating.AggregateRatingPeriod.RatingTrend
+import com.etrusted.android.trustbadge.library.model.AggregateRating.AggregateRatingPeriod
+import com.etrusted.android.trustbadge.library.model.AggregateRating.AggregateRatingDistribution
 
 internal data class ChannelInfo(
 
@@ -40,52 +37,6 @@ internal data class ChannelInfo(
     val year: AggregateRating,
     val overall: AggregateRating
 ) {
-    data class AggregateRating(
-        val count: Int,
-        val rating: Float,
-        val distribution: AggregateRatingDistribution? = null,
-        val period: AggregateRatingPeriod? = null,
-    ) {
-        data class AggregateRatingDistribution(
-            val oneStar: Int?,
-            val twoStars: Int?,
-            val threeStars: Int?,
-            val fourStars: Int?,
-            val fiveStars: Int?
-        ) {
-            companion object {
-                internal const val ONE_STAR = "oneStar"
-                internal const val TWO_STAR = "twoStars"
-                internal const val THREE_STAR = "threeStars"
-                internal const val FOUR_STAR = "fourStars"
-                internal const val FIVE_STAR = "fiveStars"
-            }
-        }
-        data class AggregateRatingPeriod(
-            val start: Date,
-            val end: Date,
-            val firstConsideredReviewSubmission: Date?,
-            val lastConsideredReviewSubmission: Date?,
-            val calculatedAt: Date?,
-            val ratingTrend: RatingTrend?
-        ) {
-            enum class RatingTrend(val raw: String) {
-                NEGATIVE("NEGATIVE"), NEUTRAL("NEUTRAL"), POSITIVE("POSITIVE");
-                companion object {
-                    fun forRaw(raw: String): RatingTrend? =
-                        values().firstOrNull { it.raw == raw  }
-                }
-            }
-            companion object {
-                internal const val START = "start"
-                internal const val END = "end"
-                internal const val FIRST_SUBMISSION = "firstConsideredReviewSubmission"
-                internal const val LAST_SUBMISSION = "lastConsideredReviewSubmission"
-                internal const val CALCULATED_AT = "calculatedAt"
-                internal const val RATING_TREND = "ratingTrend"
-            }
-        }
-    }
 
     companion object {
         private const val KEY_GRADES = "grades"
@@ -166,26 +117,4 @@ internal data class ChannelInfo(
             )
         }
     }
-}
-
-internal fun AggregateRatingDistribution.Companion.fromJson(body: JSONObject):
-        AggregateRatingDistribution {
-    return AggregateRatingDistribution(
-        oneStar = body.optInt(ONE_STAR),
-        twoStars = body.optInt(TWO_STAR),
-        threeStars = body.optInt(THREE_STAR),
-        fourStars = body.optInt(FOUR_STAR),
-        fiveStars = body.optInt(FIVE_STAR),
-    )
-}
-
-internal fun AggregateRatingPeriod.Companion.fromJson(body: JSONObject): AggregateRatingPeriod {
-    return AggregateRatingPeriod(
-        start = Date().fromString(body.getString(START)) ?: Date(),
-        end= Date().fromString(body.getString(END)) ?: Date(),
-        firstConsideredReviewSubmission = Date().fromString(body.optString(FIRST_SUBMISSION)),
-        lastConsideredReviewSubmission = Date().fromString(body.optString(LAST_SUBMISSION)),
-        calculatedAt = Date().fromString(body.optString(CALCULATED_AT)),
-        ratingTrend = RatingTrend.forRaw(body.optString(RATING_TREND)),
-    )
 }
